Create module tests with sample data file and test_spark.py with tests for app.
