+++ tests/test_file.csv
@@ -1,0 +1,16 @@
+mmsi,country
+4758772566,Australia
+2325448077,Spain
+5496043483,Greece
+5280552792,Lebanon
+3858763999,Mexico
+2652667948,Denmark
+5616746557,Turkey
+5748961587,Norway
+2890962157,Netherlands
+7905805529,Sweden
+7813304182,Poland
+2445102900,France
+2412065112,Italy
+2983392950,United Kingdom
+6313499837,United States
